Move changeset calculation to own file
